feat(cmd-action): allow running actions on leader without specifying unit number
Before this commit, `juju run` required the user to specify a unit by one of:
* full unit id, with number ("mysql/0")
* application name with leader alias ("mysql/leader")

After this commit, users can also specify only the application name and that will imply the action
should run on the leader.  For example, `juju run mysql` would resolve to `juju run mysql/leader`

A downside of this change is that, although the CLI still catches input errors, in some cases the
error message cannot be as specific as it was previously.  This is because application names,
action names, and some key-value argument keys share the same schema.  For example, if a user does
`juju run mysql/5 mysql action-name kvarg-missing-equal-sign`, we cannot definitively say
whether "kvarg-missing-equal-sign" is a malformed key-value argument, or
if "kvarg-missing-equal-sign" is the action name and instead "action-name" is actually a second
application name.  In either case, users are alerted to there being *an* error.