fix: peergrouper does not factor pending hosts
If the peergrouper sees controllers before they are provisioned and
fails to find a single cloud-local address to use for Mongo, it sets a
status message to that affect it sets a status with a message to that
effect, but assumes they are running and sets that status.

This means that the provisioner sees them, but never provisions them
because they are no longer in the pending status.

By disregarding controller hosts not yet provisioned, we allow the
provisioner to provision them after which they are correctly factored in
replica-set calculation.