feat: export charm metadata and manifest
This updates the description package to v6. The description v6
package offers the exporting of metadata and manifest files.
The 4.0 branch requires this information during the importing of
an application to ensure that RI (referential integrity) is kept
between the tables upon creation.

Unfortunately, migration between Juju versions did not export the
required information during a serialisation of the model state.
Instead, the charm information was migrated as a charm blob and
then interned into state. The reason that the charm types were
stored directly into mongo was to aid this development without
writing translation code. Warnings about this was added to the code
highlighting the problem[1].

With 4.0 we can no longer get away with this.

The solution is to pass a charm as a serialised object through
model migration. With that information we can correctly store a
charm. This does mean that if the charm does change, it is now
the responsiblity of the migration code to correctly translate
that information.

1. https://github.com/juju/juju/blob/3.6/state/charm.go#L168-L180

Signed-off-by: Simon Richardson <stickupkid@gmail.com>