chore: context aware api client package
This is the final work in a long running effort to make the api
packages context aware. Prior to this set of changes, both the
controller and agent api packages were changed to use the context
package. This moves everything across.

This unlocks a lot of potential for numerous reasons:

 1. Tracing thoughout the code base should now be possible for wiring
    up. There are some parts of the code base that is already setup
    it requires more workers to create trace roots to full complete
    this move.
 2. Log output can now include a trace id, once the logger accepts
    a context.
 3. Model loggers based on the context. Instead of having multiple
    loggers everywhere, we have one logger that can handle multiple
    models. Only on output do we write the log or not. This is
    required for status history.

The code is just one big labourious change. Adding a context.Context
to every api/client method. Fun times.