refactor: image metadata manager dqlite
This commit extends work done in the image metadata manager api facade
to facilitate a move over to dqlite. In this commit we have removed
references to mongo state for the controller uuid and also for model
metadata.

This now leaves just the image metadata still using mongo based state.