chore: description package to v8
This moves the description package to v8, which is a unreleased
version of the description package. The idea being that we want
to keep iterating on the description package without constant
releases.

Why this happened?

This happened because we need to make a breaking change to v5,
except that the main branch had already taken v6. As this is a
public repo, we shouldn't just break v6, how very unlikely may be
using it. So instead, we made a v7, 3.6 will use that branch. Main
will drop support for v6 and move to an unreleased v8. If we need
the v8, we can just move that to v9. We only then bake the v8 (or
other version) into the release, once main branch is ready. Not
before.