test: correct gomock setup in apiserver worker tests
We should not be setting up mocks inside a test suite's SetUpTest
method - this leads to problems in testing as failures on those mocks
will not be surfaced.

Refactor the apiserver worker tests so that mocks are setup inside a
setupMocks function which is called explicitly. In most cases, we don't
actually need to create the mocks, we just need a non-nil value (e.g.
for config validation).