feat: update dqlite build script
This updates to the latest dqlite release. This includes removing
the building of raft as an explicit dependency. Dqlite lib directly
builds the dependency.

This is stage 1 of the update. Stage 2 of the release is to land this,
and run the jenkins job to produce the artefacts and land the
hashes into the 3.6 branch.