chore(provisioner): move provisionertask to own package in internal
We had the provisioner task inside the provisioner worker package, and
this is in part problematic because the provisioner worker has become
somewhat overly complicated and changes on it are difficult.

This patch addresses this partly by moving the provisioner task, which
is basically the logic for building the tasks, to its own package,
leaving the worker for environ and container provisioner mainly focused
on the dispatch and orchestration instead of the logic.