feat: add controller model uuid to db and update ControllerModel
The controller model UUID was not stored anywhere in the dqlite db. The
naturual place for this to go is the controller table, alongside the
ControllerUUID.

The method on model, ControllerModel, got the controller model by
looking for the model with the name "controller". A new method on the
model state has now been added to get the controller model by looking up
its uuid in the controller table.