feat: export charm actions, config and lxd profile
This follows on from PR[1], which already exported the manifest
and metadata for a charm. This continues on with the work and exports
the charm actions and config.

Doing this allows us to construct a full charm upon importing a model
from another another controller. Thus we don't need to update any
additional information when importing the charm blob. The data already
exists.

Again, 3.6 never cares about importing charm metadata, manifest, actions
and config, so this can be elided.

 1. https://github.com/juju/juju/pull/17820