feat(application): resource ddl schema
Adds the resource DDL schema.

A resource describes a charm resource that can be used by the deployed
application.

A resource_meta is the application id and resource metadata necessary to
use a resource. Taken from the charm metadata, however once deployed,
the resource is not longer linked directly to a charm and may act
independently.

A application_resource is the resource that units of an application
should use.

A unit_resource is the resource that the given unit is currently using.

A repository_resource is a resource found by the charm revision updater
worker with the intention letting users know that a new resource revision
is available if they wish to refresh.

A resource_supplied_by defines what the value of supplied_by refers
to. There are 3 types possible. RI on the names is not required as they
can refer to removed entities and do not directly reference other
tables.