fix: return empty results
There is an issue when attemting to try set the addresses. Returning
a bare error was incorrect. Instead, we should just return an empty
set of results in this instance, to prevent changes in the 3.6 agent.

None of the original code was tested between 3.6 -> 4.0, otherwise this
would have been caught. I think we should spend time improving the
tests to ensure this doesn't happen again. Process failed us here.