feat: pass model uuid through the context
To start the work on supporting model level logging filtering we
need the model uuid to passed via the context for every request.

Do this should allow every request against a facade method call to
be assigned the right model uuid. The next change would be to ensure
that the logging package consumed a context and added the model uuid
to the associated log request. The work already exists in the
loggo package, it's just a case of wiring it up.