refactor(apiserver): wire provider factory through apiserver
Many facades currently need direct access to a provider or environ.
This is not a good pattern but we have to live with it for the moment.

To help us gradually move these facades to dqlite, we are adding a
provider factory dependency to the apiserver. This will allow us to
remove usages of EnvironConfigGetter in the facades and replace them
with a provider factory.

This is a temporary measure. Long term, any facade logic dealing with
providers/environs should be moved into the service layer. Once that is
all done, we can essentially revert this commit.