refactor(apiserver): wire provider factory through apiserver
This PR adds a provider factory dependency to the apiserver. This will
allow us to remove usages of EnvironConfigGetter in the facades and
replace them with a provider factory. Eventually, most of the facade
logic that needs providers should be pushed into the service layer. It
is possible that some facades may still need to use providers/environs
though.