feat(ports): implement opened ports state layer
After inspecting our current usage of opened ports in MongoDB, I have
come to the conclusion that the new state layer only needs two
functions.

One to open and/or close ports on a particular unit, across that unit's
endpoints, and another one to get the opened ports of a unit across it's
various endpoints.

I have done this by adding a PortRanges type into core/network, which
represents a slice of PortRange, along with a constructor + methods
which merge port ranges, ensure they're sorted, etc.

I have not yet implemented a system to reconcile the global endpoint
(represented by the empty string). This can be handled entirely within
the service layer.