feat(ports): implement opened ports state layer
After inspecting our current usage of opened ports in MongoDB, I have
come to the conclusion that the new state layer only needs two
functions.

One to open and/or close ports on a particular unit, across taht unit's
endpoints, and another one to get the opened ports of a unit across it's
various endpoints.

I have also implemented an 'IntegerInterval' construct, in
core/interval, which implements a general datastructure for an interval
of integers, along with the operation needed to combine and remove other
integer intervals.

I have not yet implemented a system to reconcile the global endpoint
(represented by the empty string). I suspect this can be handled
entirely within the service layer.