fix: ignore not implemented error
Upgrade series is no longer supported in the 4.0, so it will return
a not implemented error. As the worker is no longer required in the
4.0 land, we can return cleanly and the worker will be removed. This
is safe to do, as the migration checks that the upgrade is not in
progress.

We'll have to be careful what is removed from 4.0 once 3.6 is removed
as we won't have the privilege of providing constant hot fixes.