feat: update description package
Updates the description package, brings in the new charm actions,
config and LXD profile. This allows us to create the application
without the current work around.

Doing so means that will have all the data that we expect to keep
RI.

Work to correctly insert the charm will come as subsequent changes.