feat: import charm metadata
Imports the charm metadata from the description package into the
charm application. All the information should be provided by the
latest 3.6 migration code. This will then allow us to keep RI
for creating an application.

I've added additional information about not being tempted to make
the description package align with the internal/charm package. The
divorce of description package and the intenral/charm package is
done on purpose. This layering allows us to handle the version
negotiation and ensure that we can massage data into other data.

More work is required to remove the dependency on the internal
charm package, but this is the interim step.