feat: import charm manifest
Following on from the charm metadata, the charm manifest is imported
from the charm description package.

The charm import code currently expects that the channel is non-empty,
this might turn out to be a bit strict in the long term. It's easier
to slacken the belt, than to tighten it