feat: import charm config + actions
Import the charm config and actions from the descriptions package.
This should allow us to import a 3.6 model into a 4.0 model. Importing
a 4.0 model into a 4.0 model from another controller, isn't currently
supported as we don't yet export the charm. This will come in another
change set. Even if the data is empty we should still be able to import
a charm. A work around for importing 4.0 into 4.0 will be required
to make this happen.