feat: expose get charm
Adds a GetCharm call, which exposes all the fields of a charm at once.
This is ideal for use cases such as model migration. It is not expected
that this call will be used in other places, as it's very expensive
to get all the data at once.

It was easy to implement, because we could piggy back on existing
transaction functions and just expose them for the get charm call.