refactor: keymanager service into parts
This commit refactors the key manager service into two services. The
first service offers up the basic crud operations needed within a Juju
controller for managing a users public ssh keys. The second service
extends the crud operations of the first service adding the ability to
import ssh keys from external sources such as github and launchpad.

I have mad this change as there is multiple places that require access
to the keymanager service but not all places need the ability to import
ssh keys so the burden of having to setup a http client for the purposes
of importing can be avoided.