fix(bootstrap): fix create model version compare
Originally, when bootstrapping we would verify that the agent version
matches exactly the controller version.

This check was causing failures in our integration tests, as the build
numbers didn't match.

This check is actually way to strict. Really, we only care that the
major and the minor versions match. Amend the check.