fix: setuping mocks inside deployerSuite SetUpTest
We should not set up mocks inside a SetUpSuite or SetUpTest function.

The problem is that the *gc.C provided to the SetUpSuite or SetUpTest function is not
the same as the *gc.C provided to the actual test case. This means that if there is a
failure on the mocks (such as an unexpected or missing call), it will be reported on
the wrong *gc.C, hence it will not cause the test to fail. Doing this can cause a lot
of problems in testing which are hard to track down.