feat: export charms
Adds the ability to export charms to existing applications. As we
don't export applications directly, we need to iterate over the
existing applications and export the charm data associated with
the charm URL. Once applications are exported, we can remove the
workaround and just rely on the charm UUID.

There is a lot of work to import and export charms, but it does
allow us to correctly intercept a charm before it goes on the wire.
If the charm package changes, we'll see the broken changes at the
type level, rather than just silently ignoring them. The concept is
to make the migration steps more robust, this should provide the
foundations for that.