test: import tests
Adds application modelmigration charm import tests. Unfortunately,
the mongo state package didn't have to go through this rigmarole
when performing charm migrations. Thus we're writing now logic to
ensure that we keep RI between migrations.

The import test code ensures that we correctly handle the types
comming from a 3.6 model. We might be too strict with some of the
validations, but it ensures that there isn't any invariance in stored
values. We're strict because we need to be strict.