test: export tests
Adds charm export tests. This just covers the expected fields and
properties are correctly filled in. This is just a lot of data
shuffling between the description package and the state package.
It's potentially viable to shortcircuit that and just return the
domain types, but that's probably a lot of additional work, and one
that we can put off for now.