feat: ensure we encode action params
When we get the actions params from the description package, we end
up loosing the types that were originally set. Fundamentally it
means that the following is no longer able to be JSON marshalled:

    var m map[string]map[string]any

Ends up becoming from the description package:

    var m map[string]map[any]any

When attempting to marshal that into a blob at the application
charm service layer, the function returns:

    json: unsupported type: map[interface {}]interface {}

Even though we know that it was validated and safe to be inserted
into the database.

The fix is to recursively walk over the map and fill in the data.
The charm package prevents deeply nested action params, so it's
safe to do this recursively.