feat(application): add resource service
This is the API to interact with resources in the application domain.
Pending resources are no longer required as the application and
application resource can be added as one transaction now.

Once state resources are removed, it will be necessary to consolidate
the different resource types between the domain and core/resources as
well as incorporate the ones from internal/charm.