feat: remove fan config facade
This removes the fanconfigurer agent facade. We no longer support
fans in 4.0 (noble/24.04 has dropped support) and in doing so we've
removed the worker. This goes one step further and this removes
the facade and the model config attribute.

We'll need to fix 3.6 to ignore the "not implemented" error when the
agent is migrated. This will be done in a follow up PR once this
has landed.