refactor: remove juju-client-key as reserved
Remove the ssh public comment "juju-client-key" as reserved. When the
juju client generates an ssh key for the juju user it does so with a
comment of "juju-client-key". We special case this comment so that in
normal ssh key operations keys can't be added that have this comment.

This logic doesn't make a great deal of sense as we allow the key to be
added through bootstrap and model creation but we don't allow it via
add-ssh-key.

We also don't have any logic or security boundaries that treat these
keys differently. It doesn't make sense to have it as a reserved
keyword in the system.