feat: remove branches/generations
This removes branches/generations. This was never fully realised
and was always behind a feature flag. It was decided to remove this
and tackle it once the dqlite was done. The approach is to leave
the params/facades with the branch name and just set it to an
empty string. The underlying data/persistence model will change now
we're moving to dqlite.