refactor: remove authorized keys from addmodel
Previously when adding a new model to the controller we would support
the user specifying additional authorized keys for the model via
--config authorized-keys=... . We can know longer do this as authorized
keys is not part of model config.

On top of this if a user didn't specify any authorized keys for the
model we would attempt to load in some defaults from the users computers
without communicating this fact. What this did was introduce a security
authorisation for the model that the user wasn't explicitly aware of.

When creating a new model there are no machines available straight away
that are capable of being ssh into. It makes no sense to preload all of
this information up front.