chore(bundle): refactor bundle facade tests
Refactor apisserver bundle facade tests to ensure we do not instantiate
a gomock controller within suite SetUpTest

The reason for this is the *gc.C in SetUpTest is not the same as in the
tests themselves, which can lead to false positives tests.

Refactor to the accepted pattern, placing "defer c.setUpMocks(c).Finish()"
at the start of each test method.

We also then need to construct the facade, since it depends on our mock.