feat: add get application charm
ApplicationCharmInfo in the apiserver package requires the whole
charm package to be sent back, including the lxd profile. This is
somewhat painful, as it has no guarantees to exactly look like
the charm metadata (et al). However this is a requirement to keep
API compatability.

The code has to be moved to the commonStateBase to be able to
be read from the other service. I think potentially we should just
have one service, as the cross over is large.