chore: remove remote application icon lookup
The apiserver has the ability of serving application charm icons,
when requested. This was all well a good, except it didn't work
for cross-models aka remote applications on different controllers.
There is no redirect to point the request over to another controller
and the remote application might have been moved from there as well.

To read the icon from the charm, it has to contact the underlying
objectstore and open the charm on the controller. Then select the
charm icon (which might not even be there!) and return that back.
Not only does this feel like a waste of resources, but it would have
probably been better to just pass the icon blob with the remote
application integration.

To cement the nail in the coffin, the juju-dashboard doesn't use
this path either. Instead is favours reading the icon directly from
charmhub[1].

 1. https://github.com/canonical/juju-dashboard/blob/b3fb186bef033
    f07526dd42e5249737488fb6cd9/src/components/CharmIcon/CharmIcon
    .tsx#L12-L32