refactor: machine manager facade
This commit refactors the machine manager facade to move usages of model
config over to dqlite backed services. There still exists one usage in
this commit that will need to be cleaned up later around agent version.

Remove the need to use enviorn getter for fetching a models provider.
This facade now uses the provider tracer to extract the the two
providers out that it requires.

Removed all uses of state based model information in favour of Dqlite
services.

Updates tests to be in the same package as the facade.