feat: wire up tables for charm origin
Push charm origin through the create application method. We require
the charm origin to get the revision of the charm. Currently we
only allow the selection of a charm using the name and the revision.
This will then allow us to construct the normalized key at the api
level.
The work starts to push that concept harder.