feat: state base run transaction
Run starts to add the ability to call discrete state methods within
a transaction scope, without exposing the underlying transaction
implementation at the service layer.

It solves the problem of trying to execute business logic at the
service layer without exposing sql.Tx or sqlair.TX types. The
DBContext interface type is a typed context.Context that can be
provided as paramter to the Run closure.

The Run method gets a DB and then creates a transaction. Within
that transaction a dbContext is then created which captures both
the context and the sqlair.TX. This tuple can then be passed as
an argument to the closure. The dbContext ensures that the sqlair.TX
can only be within the closure for the lifetime of the transaction.
This is to prevent capturing the DBContext and using it outside of
the Run method. This will cause a error later on in the state
code. The tuple was deinfed in this way, as it's not possible to
modify a generic context.Value without creating a new context, or
creating a bespoke impelementation. Thus the tuple presents the
best approach of ensuring safety and easy access.

It should be noted that because of go invariance it's not possible
to swap a Run closure for the following type:

    st.Run(ctx, func(ctx context.Context) error { ... })

This will simply not compile. The Run function must be of instance
type:

   st.Run(ctx, func(dbCtx DBContext) error { ... })

Lastly, I've added a note that the DBContext implementation (dbContext)
can be pooled[1] to keep the allocations of the dbContext low. That
is not part of this changeset and can be done in future work.

  1. https://pkg.go.dev/sync#Pool