feat: execute run tx context
Now that we have a TxContext we need to be able to run it. The
package level Run method allows the unpacking of a txContext and
exposes the two parts of the tuple. It provides consistency
checks to ensure that the user isn't doing the wrong thing, and
then executes the closure.

The CoerceError at the end prevents the leaking of sql errors
to the service layer.

The Run function could be a method on the StateBase if we're looking
to track logs and metrics, but currently all the metrics and logs
are captured by the database/transaction package.