feat: add unit life watcher
Add a watcher to emit events for when the life
of an application's units changes. The semantics
are the same as for the existing mongo lifecycle watcher.

We will need to do this for other entities, so a generic
namespace watcher mapper is implemented. It just needs a
function to return the life details for the supplied
entity ids and the generic code does the rest.

For the application unit watcher, the base namespace watcher
will emit ids of all units regardless of application. So
long as the provided life getter func only returns life
values for units belonging to the specific application, the
extra ids are ignored.