chore(portranges): use PortRanges constructor/methods
Recently we replace the value type for the maptype aliased to
GroupedPortRanges from []PortRange to PortRanges

However, PortRanges was intended to be manipulated only via it's
constructor + methods, since we make certain assumptions about it's
structure.

This means constructing a PortRanges via append and the like, as we did
for values of GroupedPortRanges, could lead to bugs.

Uses these methods + constructors whenever we manipulate a
GroupedPortRanges.