feat(application): wire up RegisterCAASUnit to use run atomic The UpsertCAASUnit service method is renamed to RegisterCAASUnit. It is implemented using the new run atomic infrastructure. Various state methods are added which are used in the workflow.
The application service is composed of Charm and Application structs.
These in turn were using a single State which embedded Charm and Application state.
With the new base RunAtomic, to avoid compile errors due to ambiguous func selection,
the single State struct is no longer used and the service is created using separate
Application and Charm states.