refactor: resolve model uuid
The resolution of the model UUID is extremely flakey. If the model
UUID is an empty string then we assume that it's a controller only
login. Except that there are some locations that don't check if the
empty model UUID is correct. We can then attempt to access a state
from the state pool with an empty string. How this hasn't blown up
in prior juju versions is a minor miracle.

The solution is to push the controller only login through the helper
method: RequestModelUUID. Using this approach you can see if the
request was actually for a model UUID or not. You're now forced
to handle that method explicitly, rather than passively.