refactor: enforceble atomic context
AtomicContext composed context.Context by embedding the type, which
meant that it could be swapped out with context.Context. This turned
out to be an issue as it means if you're not careful you can end up
passing in the wrong context.Context. Even worse is that you could
run a transaction inside of another transaction.

The solution was to prevent the embedding of the context.Context and
offer it as a method.

We could go even further and make txn() a private method on the
AtomicContext interface as a marker method (only types inside of the
package can conform to the interface). With the aim of removing
runtime error. Only the atomicContext can be used as the implementation
of the interface. Thus removing the if check at runtime.
I haven't gone that far, as there is fact that people dislike that
sort of "hack" (I can see why). For now preserve the if check at
runtime and make it harder to impersonate.