chore(ports): refactor ports domain to simplify
The original implementation of the ports domain state layer included a
clever implementation to merge port ranges if they overlapped, or
subtract the different when opening and closing port ranges

This would have been a large upgrade from the current system, which does
not know how to handle overlapping port ranges. We would simply error
out in this case. The user had to deal with the exact matching port
range.

However, the new system turned out ot be a little too intricate, and may
have required a lot of work to get working on all providers. So fall
back to the original flawed system to be safe