feat: wire up apiserver common api
Wires up the apiserver common packages. This reads the charm
information from the application service. Including adding the
charm origin to the charm and application service. The charm
origin is used to store the schema/soruce and the revision of the
charm. Using that information it's possible to reconstruct the
charm URL. The architecture is still missing from the charm URL,
but we'll have to see if we actually require that to exposed
at the client side. If it does need exposing then, we'll just
need to push that into the table row when we're dual writing.

Lastly, I updated the common/charms package test to ensure that
we can full write a internal/charm to a params charm. Unfortunately,
the action results wasn't correct, so if anything needed that
it was missed. We're now exporting if it's available. This is
reflected in the schema.json file