feat: add everyone@external user on bootstrap
The everyone@external user is used to grant permissions to all external
users in the database. In mongo this user didn't need to exist as the
user that permissions were granted to didn't need to exist.

Now that we have referential integrity this user needs to exist to grant
it permissions.

When an offer is added the everyone@external user needs
to be granted read permissions on it. This means it needs to already
exist in the database. It is now added by the bootstrap worker.