fix: double writing external user
fix the error where on login we would attempt to add an external user to
the database even if it already existed. As part of the fix refactor the
function into a seperate function to ensure that the missing
user is added. The ReadUserAccessLevelForTarget can then be used after
that to actually check the users permissions.