feat: expose platform and origin correctly
The application AND charm need their own platform. Although they
share the same data, they can drift over time. This shouldn't happen
and the likely only place is because of the force flag, even so
we need to model it out juju works atm.

The problem is we now lose the property that applications are just
manifestations of charms. This is not actually true any more. Instead
they're an instance of a charm, with their own properties on top.
This isn't ideal as it's then not possible to ensure strict
representation of each.

The solution is to remove the views and just have hardcoded tables.