refactor: redefine function signature for RequestModelUUID and others
RequestModelUUID does not need an http.Request, it only needs a context.
Supply what is needed not more.

Other methods for an httpContext only need a context as well:
stateForRequestUnauthenticated, objectStoreForRequest,
controllerObjectStoreForRequest, and serviceFactoryForRequest.

Some methods of httpContext look like they only required a context as
well, however they must share a method signture with a different method
as they are supplied to structures used twice with different methods.
E.G. methods which must satisfy the stateAuthRunc for the toolsHandler
structure.