test: ensure charm origin is correct
The domain package now ensures that types that are passed in the
creation of the application/charm must have valid data. We're no
longer accepting partial data for parts of the charm that we MUST
know about. The architecture and the charm source is a prime
example. These MUST be known.