feat(cmd-action): enforce that `juju run` uses units of a single application
Before this commit, `juju run unit-name [unit-name...]` did not validate whether all units in
the call were of the same application.  This was unintentional.

This commit adds validation in the juju CLI to enforce that all units must be of the same
application, as well as clarifies this in the help text.