refactor: rename CancelMachineReboot method to ClearMachineReboot
- Before this commit, the method for clearing a machine reboot flag was named CancelMachineReboot.
- After this commit, the method for clearing a machine reboot flag is renamed to ClearMachineReboot
  for better clarity.

The idea here is that to mimic the old behavior, there is no such thing as "cancelling" a reboot.
Requesting a reboot will trigger a notify event for a machine (as a side effect of adding the
machine UUID to the machine_requires_reboot table), and such request cannot be cancelled.
The functionality of removing the uuid is a clearance, since it doesn't prevent a reboot, but
instead allows a reboot to be requested again.

Signed-off-by: Guillaume Fouillet <guillaume.fouillet@canonical.com>