feat: save the charm reference name
The reference name is the charm name that is associated with
the application. To ensure that we can always locate the charm by
name because of existing APIs we need to be able to get it from
just the name with the revision.

This can happen if the charmhub name differs from the charm name
in the metadata and we can't rewire the charm name. This ensures
backwards compatibility between clients at the cost of complexity
at the service layer.