fix: model last login flaky test
The model last login test called time.Now twice and assumed that they
would be at least a microsecond apart. This is not the case on a certain
someones shiney new laptop.

Change the UpdateModelLastLogin function to take an explict time value
to enter into the database. The service function calls time.Now. (This is
also useful for migration as we can set old model last logins now).

Change the test to use time values that are explicty set and checked
against rather than relying on the differece between calls to time.Now.