refactor: replace model config calls in Controller facade
This PR removes usages of Mongo model config calls from the Controller
facade. They are replaced with calls to a model config or model info
service as appropriate.

The HostedModelConfig method needs to access the model config for every
model in the controller. Hence, we need to convert this facade into a
multi-model facade, so that it can access the model service factory for
an arbitrary model.