chore: split provisioner worker in machine and container provisioners
This patch splits the provisioner worker into two new workers: machine
and container provisioners.
This follows a previous patch (https://github.com/juju/juju/pull/17954)
on which we moved the provisioner task to its own package. The goal of
all this work is to facilitate the refactorings and trims on the
provisioner which was a huge worker.