fix(config): allow un-setting ha-space controller config
Before this patch it was impossible to un-set the juju-ha-space
controller config value because:
- first, the empty string was rejected because it does not match the
  regexp on the names package for space names.
- second, because on the state layer we check that the provided space
  exists, and an empty string space never does.

The first reason was fixed on `controller/config.go` by simply adding
the "not empty string" check before validating the space name.

For the second reason, a workaround had to be added: if the passed space
name is the empty string, then we don't check for its existence. This is
not good because in the `UpdateControllerConfig()` method we already
have the `removeAttrs` argument but on the only usage of this method (on
the client facades) we pass a nil value as `removeAttrs`.