feat(ports): implement ports service layer
Implement a service for ports corresponding to the state interface
UnitPortRanges.

There is a singular massive complication which prevents this service
from simply being a shim around state. That is the behaviour of the
wildcard endpoint.

I have laid out the current behaviour in part of this
[spec](docs.google.com/document/d/1TbkiOZ760Om1xoKolQWOv_UVd7-NOsji915gcIQmsmw)

Essentially, the wildcard endpoint represents a short-hand for all
endpoints. So what we apply to the wildcard overrides all other
endpoints. This means some fiddly manipulation of the opening + closing
grouped port ranges.

To do this properly, we also need to make some extra calls to state.
Specifically, GetEndpoints and GetEndpointOpenedPorts. Implement these
new state functions.

NOTE: GetEndpoints is currently only a best guess. To the full correct
list of endpoints, we need to wait until the charm domain has been
implemented.