refactor: migration target facade
This commit removes version 1 and 2 from the migration target facade so
that in Juju 4.0 version 3 of the facade is the only supported version.
This version is the current version supported by 3.6.

We have removed the older facade versions as we need to swap the
migration target facade over to a multi model facade context as it
operates on not only the controller models but the models that are also
being migrated to the target controller.

This can be seen in calls such as CheckMachines.