feat: modelmigration service
This commit introduces a new model migration service to the set of
domains we have in Juju 4.0. This new domain is here to answer model
migration questions on behalf of the controller.

Included in this new service is a check for verifying machines after
model migration between what the Juju database has and what the cloud
reports.